Remove old tests; stop completer from crashing on manpage parsing; add global directive to set global variables; only print output when a command actually returns it (don't print None)
